reorganize Base with default packages, installed in site/
as a demo, moves DataFmt out of Base. adds a line to default juliarc
so the functions are still available by default.
