+++ base/sysimg.jl
@@ -360,8 +360,6 @@ include("mmap.jl")
 import .Mmap
 
 # utilities - timing, help, edit
-include("datafmt.jl")
-using .DataFmt
 include("deepcopy.jl")
 include("interactiveutil.jl")
 include("summarysize.jl")
@@ -452,5 +450,13 @@ include(Base, "precompile.jl")
 end # baremodule Base
 
 using Base
+
+# set up load path to be able to find stdlib packages
+Base.init_load_path(ccall(:jl_get_julia_home, Any, ()))
+
+# load some stdlib packages but don't put their names in Main
+Base.require(:DelimitedFiles)
+
+empty!(LOAD_PATH)
 
 Base.isfile("userimg.jl") && Base.include(Main, "userimg.jl")
