+++ doc/src/stdlib/.gitignore
@@ -1,0 +1,1 @@
+delimitedfiles.md
