+++ etc/juliarc.jl
@@ -1,3 +1,4 @@
 # This file should contain site-specific commands to be executed on Julia startup
 # Users may store their own personal commands in the user home directory `homedir()`, in a file named `.juliarc.jl`
 
+using DelimitedFiles
