+++ stdlib/DelimitedFiles/docs/src/index.md
@@ -1,0 +1,11 @@
+# Delimited Files
+
+```@docs
+DelimitedFiles.readdlm(::Any, ::Char, ::Type, ::Char)
+DelimitedFiles.readdlm(::Any, ::Char, ::Char)
+DelimitedFiles.readdlm(::Any, ::Char, ::Type)
+DelimitedFiles.readdlm(::Any, ::Char)
+DelimitedFiles.readdlm(::Any, ::Type)
+DelimitedFiles.readdlm(::Any)
+DelimitedFiles.writedlm
+```
