+++ stdlib/DelimitedFiles/src/DelimitedFiles.jl
@@ -1,12 +1,16 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-## file formats ##
+__precompile__(true)
 
-module DataFmt
+module DelimitedFiles
 
 import Base: _default_delims, tryparse_internal, show
 
-export countlines, readdlm, writedlm
+export readdlm, writedlm
+
+Base.@deprecate readcsv(io; opts...) readdlm(io, ','; opts...)
+Base.@deprecate readcsv(io, T::Type; opts...) readdlm(io, ',', T; opts...)
+Base.@deprecate writecsv(io, a; opts...) writedlm(io, a, ','; opts...)
 
 invalid_dlm(::Type{Char})   = reinterpret(Char, 0xfffffffe)
 invalid_dlm(::Type{UInt8})  = 0xfe
@@ -14,29 +18,6 @@ invalid_dlm(::Type{UInt16}) = 0xfffe
 invalid_dlm(::Type{UInt32}) = 0xfffffffe
 
 const offs_chunk_size = 5000
-
-countlines(f::AbstractString, eol::Char='\n') = open(io->countlines(io,eol), f)::Int
-
-"""
-    countlines(io::IO, eol::Char='\\n')
-
-Read `io` until the end of the stream/file and count the number of lines. To specify a file
-pass the filename as the first argument. EOL markers other than `'\\n'` are supported by
-passing them as the second argument.
-"""
-function countlines(io::IO, eol::Char='\n')
-    isascii(eol) || throw(ArgumentError("only ASCII line terminators are supported"))
-    aeol = UInt8(eol)
-    a = Vector{UInt8}(8192)
-    nl = 0
-    while !eof(io)
-        nb = readbytes!(io, a)
-        @simd for i=1:nb
-            @inbounds nl += a[i] == aeol
-        end
-    end
-    nl
-end
 
 """
     readdlm(source, T::Type; options...)
@@ -698,4 +679,4 @@ writedlm(io, a; opts...) = writedlm(io, a, '\t'; opts...)
 show(io::IO, ::MIME"text/csv", a) = writedlm(io, a, ',')
 show(io::IO, ::MIME"text/tab-separated-values", a) = writedlm(io, a, '\t')
 
-end # module DataFmt
+end # module DelimitedFiles
